+++ penguins/acutis/__init__.py
@@ -6,6 +6,7 @@ Importing this module adds additional methods to Polars Series and DataFrame obj
 
 from penguins.acutis.affiche import affiche             # Imports both the function and the method 
 from penguins.acutis.count_table import count_table
+from penguins.acutis.count_null import count_null
 from penguins.acutis.pasteurize import pasteurize
 
-__all__ = ['affiche', 'count_table', 'pasteurize']
+__all__ = ['affiche', 'count_table', 'count_null', 'pasteurize']